adds 2019 day 9

+++ src/main/kotlin/aoc2019/Day2.kt
@@ -2,24 +2,24 @@ package aoc2019
 
 private object Day2 {
 
-    private const val TARGET = 19690720
+    private const val TARGET = 19690720L
 
-    private fun processWithInitialSettings(codes: MutableList<Int>, a: Int = 0, b: Int = 0): Int {
+    private fun processWithInitialSettings(codes: MutableList<Long>, a: Long = 0L, b: Long = 0L): Long {
         codes[1] = a
         codes[2] = b
         IntCodeProcessor(codes).run()
         return codes[0]
     }
 
-    fun findValueAtPositionZero(): Int {
+    fun findValueAtPositionZero(): Long {
         val codes = loadIntCodeInstructions("inputs/2019/2.txt")
         return processWithInitialSettings(codes, 12, 2)
     }
 
-    fun findInputsForTargetValue(): Int {
+    fun findInputsForTargetValue(): Long {
         val codes = loadIntCodeInstructions("inputs/2019/2.txt")
-        for (i in 0 .. 99) {
-            for (j in 0 .. 99) {
+        for (i in 0L .. 99L) {
+            for (j in 0L .. 99L) {
                 val result = processWithInitialSettings(codes.toMutableList(), i, j)
                 if (result == TARGET) {
                     return 100 * i + j
